+++ docs/release-notes.md
@@ -9,7 +9,7 @@ title: jqwik Release Notes
 <!-- DON'T EDIT THIS SECTION, INSTEAD RE-RUN doctoc TO UPDATE -->
 ### Table of Contents  
 
-- [1.3.5](#135-snapshot)
+- [1.3.5](#135)
 - [1.3.4](#134)
 - [1.3.3](#133)
 - [1.3.2](#132)
@@ -23,7 +23,14 @@ title: jqwik Release Notes
 
 <!-- END doctoc generated TOC please keep comment here to allow auto update -->
 
-## 1.3.5-SNAPSHOT
+## 1.3.5
+
+<p style="padding-left:1em;font-size:larger">
+<a href="/docs/1.3.5/user-guide.html">User guide</a>
+and
+<a href="/docs/1.3.5/javadoc/index.html">javadoc</a>
+</p>
+
 
 #### New and Enhanced Features
 
@@ -49,6 +56,8 @@ title: jqwik Release Notes
 #### Bug Fixes
 
 - edge cases generation will be stopped when 10000 edge cases have been found.
+  See https://github.com/jlink/jqwik/issues/113
+
 
 ## 1.3.4